Partially reorganized docs
- api is now split into 3 parts
- files are reorganized in folders

Additionally, some class members (Quantity, Unit, Measurement, Group, Context, System)
are excluded from the docs as they confuse sphinx.
