+++ docs/advanced/currencies.rst
