+++ docs/advanced/custom-registry-class.rst
