+++ docs/advanced/index.rst
@@ -1,0 +1,17 @@
+Advanced Guides
+===============
+
+Pint contains some useful and fun feature. You will find them here.
+
+.. toctree::
+    :maxdepth: 2
+    :hidden:
+
+    performance
+    serialization
+    defining
+    wrapping
+    measurement
+    pitheorem
+    currencies
+    custom-registry-class
