+++ docs/advanced/performance.rst
