+++ docs/advanced/pitheorem.rst
