+++ docs/advanced/serialization.rst
@@ -1,5 +1,3 @@
-.. _serialization:
-
 
 Serialization
 =============
