--- docs/advanced_guides.rst
@@ -1,15 +1,0 @@
-Advanced Guides
-===============
-
-.. toctree::
-    :maxdepth: 2
-    :hidden:
-
-    user/performance
-    user/serialization
-    user/defining
-    user/wrapping
-    user/measurement
-    user/pitheorem
-    user/currencies
-    user/custom-registry-class
