--- docs/api.rst
@@ -1,71 +1,0 @@
-===================
-API reference
-===================
-
-.. currentmodule:: pint
-
-.. automodule:: pint.babel_names
-    :members:
-
-.. automodule:: pint.compat
-    :members:
-
-.. automodule:: pint.converters
-    :members:
-
-.. automodule:: pint.definitions
-    :members:
-
-.. automodule:: pint.errors
-    :members:
-
-.. automodule:: pint.formatting
-    :members:
-
-.. automodule:: pint.matplotlib
-    :members:
-
-.. automodule:: pint.pint_eval
-    :members:
-
-.. automodule:: pint.registry
-    :members:
-
-.. automodule:: pint.registry_helpers
-    :members:
-
-.. automodule:: pint.testing
-    :members:
-
-.. automodule:: pint.util
-    :members:
-
-Facets
-------
-
-.. automodule:: pint.facets.plain
-    :members:
-
-.. automodule:: pint.facets.nonmultiplicative
-    :members:
-
-.. automodule:: pint.facets.formatting
-    :members:
-
-.. automodule:: pint.facets.numpy
-    :members:
-
-.. automodule:: pint.facets.dask
-    :members:
-
-.. automodule:: pint.facets.measurement
-    :members:
-
-.. automodule:: pint.facets.group
-    :members:
-
-.. automodule:: pint.facets.system
-    :members:
-
-.. automodule:: pint.facets.context
-    :members:
