+++ docs/api/common.rst
@@ -1,0 +1,40 @@
+Common API
+==========
+
+.. automodule:: pint.babel_names
+    :members:
+
+.. automodule:: pint.compat
+    :members:
+
+.. automodule:: pint.converters
+    :members:
+
+.. automodule:: pint.definitions
+    :members:
+
+.. automodule:: pint.errors
+    :members:
+
+.. automodule:: pint.formatting
+    :members:
+
+.. automodule:: pint.matplotlib
+    :members:
+
+.. automodule:: pint.pint_eval
+    :members:
+
+.. automodule:: pint.registry
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System, UnitRegistry
+
+.. automodule:: pint.registry_helpers
+    :members:
+
+.. automodule:: pint.testing
+    :members:
+
+.. automodule:: pint.util
+    :members:
+    :exclude-members: Unit
