+++ docs/api/facets.rst
@@ -1,0 +1,39 @@
+API facets reference
+====================
+
+
+.. automodule:: pint.facets.plain
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.nonmultiplicative
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.formatting
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.numpy
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.dask
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.measurement
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.group
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.system
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
+
+.. automodule:: pint.facets.context
+    :members:
+    :exclude-members: Quantity, Unit, Measurement, Group, Context, System
