+++ docs/api/index.rst
@@ -1,0 +1,10 @@
+API reference
+==============
+
+.. toctree::
+    :maxdepth: 1
+    :hidden:
+
+    base
+    common
+    facets
