+++ docs/user/formatting.rst
@@ -1,4 +1,3 @@
-.. _formatting:
 .. currentmodule:: pint
 
 
@@ -8,8 +7,9 @@
    import pint
 
 
-String formatting
-=================
+String formatting specification
+===============================
+
 The conversion of :py:class:`Unit` and :py:class:`Quantity` objects to strings (e.g.
 through the :py:class:`str` builtin or f-strings) can be customized using :ref:`format
 specifications <formatspec>`. The basic format is:
